feat: Add secret definitions for sample-users
This commit adds the secret definitions for the `sample-users` secret in the `secrets` directory. The secret-definition.yml file specifies the data for two users, USER_A and USER_B, encoded in base64. These additions enable the usage of the `sample-users` secret in the cluster, providing sensitive information to the `pod-secret` pod.